Cleaning up some bugs/problems.
freeboard.cfg
See FreeboardServerNotes - FreeboardUseNotes. Hand edited freeboard.cfg and layers.js to get the proper starting conditions for the list of charts. Then used freeboard to add a chart.

Ports to scan are setup for PC. position of charts has been selected to maximize display on ipad.

anchor.js
Unable to get AnchorWindow to work properly. The steelseries displays stubbornly refuse to show the values sent to them. I want to check this condition again developing on Linix.

Layers.js
See FreeboardServerNotes

pom.xml
Had to backup to purejavacomm version 0.0.17 as the latest version was giving me Termios errors on my laptop.

AnchorViwModel.java
Corrected two bugs: anchorAlarmOn was mishandled in setAnchorAlarmState and the @Listen("on click... has to refer to a toolbarbutton, not a button.

pom.xml