Update Raspberry Pi install script for Jessie
This patch automates installation of freeboard-server to raspbian
jessie, the latest release of the raspberry pi operating system.
It takes advantage of the new systemd init system which addresses
issue #2 and simplifies logging of the server output.

It has been tested in the following environment:

  * Raspberry Pi 3 Model B
  * 32G class 10 micro SD Card
  * Base install of Raspbian Jessie LITE

Fully installed, with the WiFi option, it's around 1.2G, leaving about
28G free on my SD card for charts, tracks and logs and such.

The 'Lite' version of raspbian also has less software running in the
background, so it should leave more memory and processing power for
doing real work.

Documentation has been updated in the format of a HOWTO, and also
includes some basics of console administration.  An intro to working
with systemd services is included.

systemd manages non-forking services like the freeboard-server in a
much better way than old-school init systems did (this addresses
issue #2.)

Some other features of the setup script are:

* Designed to be run at any time to re-configure and update
* Efficient (changes what it has to)
* Smart (stops, starts, or restarts services when needed)
* Will install map packages

Some things left unfinished (Still TODO)

* allow user to override default network (and possibly other) settings
at execution time. save overrides and use them as defaults on the next
execution
* allow updating of maps by comparing timestamps on the zip files
against the installed content
* validate wifi devices work with hostapd

fixes #2