logg.zul Added labels headWotSize, logSiz, latLonSiz to hold size of elements of the logg instrument
WindViewModel.java
Changed the initialization of scale from 0.5 to 1.0

doubleLog.zul
Removed the definition of the attribute onAfterSize. I have found a better way to handle remembering the instrument size.
Added the labels doubleLogWidth and doubleLogHeight to hold the size of the doubleLog instrument.

SparklineDepthViewModel.java
This is my attempt to include a sparkline in the depth instrument display on the chartplotter. It doesn't work.

util.java
Removed the line that set the DEPTH_SCALE

DepthViewModel.java
Removed support for a sparkline. I could not get it to work.
In doAfterCompose, ensured that the scale is set to 1 if it is not in the properties. 

depth.js
Changed the way the instrument size is initialized. If it is the first time and the instrument parameters are  not in freeboard.cfg, the size is taken from the labels in the zul file and the scale = 1. If it is not the first time,  the size is set to the label values*scale.

doubleLog.js
See the discussion for depth.js

wind2.js
See the discussion for depth.js

LoggViewmodel.java
Added the wire for headWptSize, logSize and latLonSiz.
Set scale = 1 if the key is not the the properties freeboard.cfg

NEMAProcessor.java
Added comments to clarify use of random gaussian for testing Depth instrument

sparklineDepth.js
Attempt to add sparkline to Depth instrument. It does not work.

SerialPortReader.java
Added a hack (now commented out) to enable replacing GGA sentences with DBT sentences for testing the Depth instrument code when only a GPS is connected to the system.

logg.js
Added variables to track the height and width of elements of the logg instrument.
The labels in logg.zul that are used to hold the sizes of the elements of the logg instrument have values that encode the width and height as wwwxhhh where www is the width and hhh is the height, the elements separated by x. The code here splits the value to recover the sizes. Otherwise the maintenance of the instrument size is as in depth.js