WIP. DoubleLog instrument is now properly placed in size and position after browser reload. The code needs to be cleaned up.
NMEAProcessor.java
Hack is in place to substitute random gaussian values for depth

SerialPortReader.java
Hack is in place to replace GGA sentences with DBT sentences

doubleLog.js
Changes to implement use of localStorage for scale, width, height and position. This needs cleaning up.

DoubleLogModel.java
Removed most of the code in this class. Have inserted a call to a js function onLoad to set the window position in doAfterCompose

ConfigViewModel.java
Removed the code that dealt with sparklinePoints. This parameter now only exists in freeboard.cfg. It can no longer be set via the Config window.

configWindow.zul
Removed the textbox line for sparklinePoints.