Debugged ToggleWind instrument, added BoatCircles, updated Steelseries.js
Debugged ToggleWind instrument using motu.log as input data.
Added Boat Circles to the Chartplotter. The number of circles and their diameter
are set in the Config window.

wind2.zul
Changed the name of the window to Wind T+Added

configWindow.zul
Corrected tooltip to tooltiptext.
Added Number of Boat Circles and Boat Circle Radii textboxes to the window.

toggleWind.js
Many changes and additions to support maintaining the wind direction and wind velocity 
arrays in localStorage. Changed the Steelseries definitions to handle displaying the 
true averages and not the previous values.

Util.java
Added code to initialize the number of boat circles and boat circle radii from freeboard.cfg
Changed the code in checkTime to do nothing if it is a Windows machine and to set the system 
time from the GPS if it is otherwise. This code has not yet been tested.

floating.zul
Added numCircles and radiiCircles as Labels

toggleWind.zul
Changed window title to Wind T/A

Constants.java
Added NUM_BOAT_CIRCLES and RADII_BOAT_CIRCLES as static final String

ConfigViewModel.java
Added code to support the numBoatCircles and radiiBoatCircles text boxes.
Changed portNAmesRegexLinix to add /dev/pts so that socat can be used to connect to a log file.

Steelseries.js
Upgraded Steelseries.js to Ver 0.14.17 from Ver 0.13.0.
Have not changed the minified version of the file.

chartplotter.js
Added code to generate the boat circles. Added a layer group.

ChartplotterViewModel.java
Added code to support the generation of the boat circles.